+++ src/driver/call_provers.mli
@@ -107,7 +107,7 @@ type prover_result_parser = {
 type prover_call
 (** Type that represents a single prover run *)
 
-type resource_limit = {
+type resource_limits = {
   limit_time  : float;
   limit_mem   : int;
   limit_steps : int;
@@ -115,33 +115,21 @@ type resource_limit = {
 (* represents the three ways a prover run can be limited: in time, memory
    and/or steps *)
 
-val empty_limit : resource_limit
+val empty_limits : resource_limits
 (* the limit object which imposes no limits. Use this object to impose no
    limits, but also to know if some concrete time, steps or memlimit actually
    means "no limit" *)
 
-val limit_max : resource_limit -> resource_limit -> resource_limit
+val limit_max : resource_limits -> resource_limits -> resource_limits
 (* return the limit object whose components represent the maximum of the
    corresponding components of the arguments *)
 
 val call_editor : config:Whyconf.main -> command:string -> string -> prover_call
 
-(* internal use only
-val call_on_file :
-  command         : string ->
-  libdir          : string ->
-  datadir         : string ->
-  limit           : resource_limit ->
-  res_parser      : prover_result_parser ->
-  printer_mapping : Printer.printer_mapping ->
-  ?inplace        : bool ->
-  string -> prover_call
- *)
-
 val call_on_buffer :
   command         : string ->
   config          : Whyconf.main ->
-  limit           : resource_limit ->
+  limits          : resource_limits ->
   res_parser      : prover_result_parser ->
   filename        : string ->
   get_model       : Printer.printing_info option ->