+++ src/session/controller_itp.ml
@@ -299,7 +299,7 @@ type sched_pa_rec =
   { spa_contr    : controller;
     spa_id       : proofNodeID;
     spa_pr       : Whyconf.prover;
-    spa_limit    : Call_provers.resource_limit;
+    spa_limits   : Call_provers.resource_limits;
     spa_pr_scr   : string option;
     spa_callback : (proof_attempt_status -> unit);
     spa_ores     : Call_provers.prover_result option
@@ -412,10 +412,10 @@ let fuzzy_proof_time nres ores =
 
 let build_prover_call spa =
   let c = spa.spa_contr in
-  let limit = spa.spa_limit in
+  let limits = spa.spa_limits in
   let (config_pr,driver) = Hprover.find c.controller_provers spa.spa_pr in
   let with_steps =
-    Call_provers.(limit.limit_steps > empty_limit.limit_steps) in
+    Call_provers.(limits.limit_steps > empty_limits.limit_steps) in
   let command =
     Whyconf.get_complete_command config_pr ~with_steps in
   match Session_itp.get_task c.controller_session spa.spa_id with
@@ -429,7 +429,7 @@ let build_prover_call spa =
     let main = Whyconf.get_main c.controller_config in
     try
       let call = Driver.prove_task ?old:spa.spa_pr_scr ~inplace ~command
-          ~limit ~interactive ~config:main driver task in
+          ~limits ~interactive ~config:main driver task in
       let pa =
         { tp_id       = spa.spa_id;
           tp_callback = spa.spa_callback;
@@ -605,7 +605,7 @@ let run_idle_handler () =
       S.timeout ~ms:default_delay_ms timeout_handler;
     end
 
-let schedule_proof_attempt ?proof_script_filename c id pr ~limit ~callback ~notification =
+let schedule_proof_attempt ?proof_script_filename c id pr ~limits ~callback ~notification =
   let ses = c.controller_session in
   Debug.dprintf debug_sched "schedule_proof_attempt called@.";
   let callback panid s =
@@ -635,7 +635,7 @@ let schedule_proof_attempt ?proof_script_filename c id pr ~limit ~callback ~noti
     end;
     callback panid s
   in
-  let adaptlimit,ores,proof_script =
+  let adaptlimits,ores,proof_script =
     try
       let h = get_proof_attempt_ids ses id in
       let pa = Hprover.find h pr in
@@ -643,8 +643,8 @@ let schedule_proof_attempt ?proof_script_filename c id pr ~limit ~callback ~noti
       let old_res = a.proof_state in
       let config_pr,_ = Hprover.find c.controller_provers pr in
       let interactive = config_pr.Whyconf.interactive in
-      let use_steps = Call_provers.(limit.limit_steps <> empty_limit.limit_steps) in
-      let limit = adapt_limits ~interactive ~use_steps limit a in
+      let use_steps = Call_provers.(limits.limit_steps <> empty_limits.limit_steps) in
+      let limits = adapt_limits ~interactive ~use_steps limits a in
       let script =
         if proof_script_filename = None then
           Option.map (fun s ->
@@ -654,15 +654,15 @@ let schedule_proof_attempt ?proof_script_filename c id pr ~limit ~callback ~noti
           else
             proof_script_filename
       in
-      limit, old_res, script
-    with Not_found | Session_itp.BadID -> limit,None,proof_script_filename
+      limits, old_res, script
+    with Not_found | Session_itp.BadID -> limits,None,proof_script_filename
   in
-  let panid = graft_proof_attempt ~limit ses id pr in
+  let panid = graft_proof_attempt ~limits ses id pr in
   let spa =
     { spa_contr    = c;
       spa_id       = id;
       spa_pr       = pr;
-      spa_limit    = adaptlimit;
+      spa_limits   = adaptlimits;
       spa_pr_scr   = proof_script;
       spa_callback = callback panid;
       spa_ores     = ores } in
@@ -714,8 +714,8 @@ let prepare_edition c ?file pn pr ~notification =
         | Some f -> f
         | None -> create_file_rel_path c pr pn
       in
-      let limit = Call_provers.empty_limit in
-      graft_proof_attempt session pn pr ~file ~limit
+      let limits = Call_provers.empty_limits in
+      graft_proof_attempt session pn pr ~file ~limits
   in
   update_goal_node notification session pn;
   let pa = get_proof_attempt_node session panid in
@@ -841,12 +841,12 @@ let call_one_prover c (p, timelimit, memlimit, steplimit) ~callback ~notificatio
   let timelimit = Option.value ~default:(Whyconf.timelimit main) timelimit in
   let memlimit = Option.value ~default:(Whyconf.memlimit main) memlimit in
   let steplimit = Option.value ~default:0 steplimit in
-  let limit = {
+  let limits = {
     Call_provers.limit_time = timelimit;
     limit_mem  = memlimit;
     limit_steps = steplimit;
   } in
-  schedule_proof_attempt c g p ~limit ~callback ~notification
+  schedule_proof_attempt c g p ~limits ~callback ~notification
 
 let run_strategy_on_goal
     c id strat ~callback_pa ~callback_tr ~callback ~notification ~removed =
@@ -1014,8 +1014,8 @@ callback STShalt
 let schedule_pa_with_same_arguments
     c (pa: proof_attempt_node) (pn: proofNodeID) ~callback ~notification =
   let prover = pa.prover in
-  let limit = pa.limit in
-  schedule_proof_attempt c pn prover ~limit ~callback ~notification
+  let limits = pa.limits in
+  schedule_proof_attempt c pn prover ~limits ~callback ~notification
 
 let schedule_tr_with_same_arguments
     c (tr: transID) (pn: proofNodeID) ~callback ~notification =
@@ -1201,7 +1201,7 @@ let find_prover notification c goal_id pr =
         end
 
 
-let replay_proof_attempt c pr limit (parid: proofNodeID) id ~callback ~notification =
+let replay_proof_attempt c pr limits (parid: proofNodeID) id ~callback ~notification =
   (* The replay can be done on a different machine so we need
      to check more things before giving the attempt to the scheduler *)
   match find_prover notification c parid pr with
@@ -1214,7 +1214,7 @@ let replay_proof_attempt c pr limit (parid: proofNodeID) id ~callback ~notificat
      try
        if pr' <> pr then callback id (UpgradeProver pr');
        let _ = get_task c.controller_session parid in
-       schedule_proof_attempt c parid pr' ~limit ~callback ~notification
+       schedule_proof_attempt c parid pr' ~limits ~callback ~notification
      with Not_found ->
        callback id Detached
 
@@ -1249,7 +1249,7 @@ let print_report fmt (r: report) =
       (Call_provers.print_prover_result ~json:false) new_r
 
 (* TODO to be removed when we have a better way to print *)
-let replay_print fmt (lr: (proofNodeID * Whyconf.prover * Call_provers.resource_limit * report) list) =
+let replay_print fmt (lr: (proofNodeID * Whyconf.prover * Call_provers.resource_limits * report) list) =
   let pp_elem fmt (id, pr, rl, report) =
     fprintf fmt "ProofNodeID: %d, Prover: %a, Timelimit?: %.2f, Result: @[<h>%a@]"
       (Obj.magic id) Whyconf.print_prover pr
@@ -1325,7 +1325,7 @@ let replay ~valid_only ~obsolete_only ?(use_steps=false) ?(filter=fun _ -> true)
         let parid = pa.parent in
         let pr = pa.prover in
         (* If use_steps, we give only steps as a limit *)
-        let step_limit = Call_provers.(empty_limit.limit_steps) in
+        let step_limit = Call_provers.(empty_limits.limit_steps) in
         let step_limit =
           if use_steps then
             match pa.proof_state with
@@ -1333,10 +1333,10 @@ let replay ~valid_only ~obsolete_only ?(use_steps=false) ?(filter=fun _ -> true)
             | Some r -> r.Call_provers.pr_steps
           else step_limit
         in
-        let limit = Call_provers.{pa.limit with limit_steps = step_limit } in
-        replay_proof_attempt c pr limit parid id
+        let limits = Call_provers.{pa.limits with limit_steps = step_limit } in
+        replay_proof_attempt c pr limits parid id
                              ~callback:(fun id s ->
-                                        craft_report s parid limit pa;
+                                        craft_report s parid limits pa;
                                         callback id s;
                                         if !count = 0 then
                                           final_callback !found_upgraded_prover !report)
@@ -1392,11 +1392,10 @@ let bisect_proof_attempt ~callback_tr ~callback_pa ~notification ~removed c pa_i
     raise (CannotRunBisectionOn pa_id); (* proof attempt should be valid *)
   let goal_id = pa.parent in
   let prover = pa.prover in
-  let limit = { pa.limit with
-                Call_provers.limit_steps =
-                  Call_provers.empty_limit.Call_provers.limit_steps }
+  let limits = Call_provers.{ pa.limits with
+                limit_steps = empty_limits.limit_steps }
   in
-  let timelimit = ref limit.Call_provers.limit_time in
+  let timelimit = ref limits.Call_provers.limit_time in
   let set_timelimit res =
     timelimit := res.Call_provers.pr_time in
   let bisect_end rem =
@@ -1419,7 +1418,7 @@ let bisect_proof_attempt ~callback_tr ~callback_pa ~notification ~removed c pa_i
           | TSdone trid ->
              match get_sub_tasks ses trid with
              | [pn] ->
-                let limit = { limit with Call_provers.limit_time = !timelimit; } in
+                let limits = { limits with Call_provers.limit_time = !timelimit; } in
                 let callback paid st =
                   callback_pa paid st;
                   begin match st with
@@ -1436,7 +1435,7 @@ let bisect_proof_attempt ~callback_tr ~callback_pa ~notification ~removed c pa_i
                        (Call_provers.print_prover_result ~json:false) res
                   end
                 in
-                schedule_proof_attempt c pn prover ~limit ~callback ~notification
+                schedule_proof_attempt c pn prover ~limits ~callback ~notification
              | _ -> assert false
           end
         in
@@ -1483,7 +1482,7 @@ later on. We do has if proof fails. *)
          callback_tr "remove" [rem] st;
          match get_sub_tasks ses trid with
          | [pn] ->
-            let limit = { limit with Call_provers.limit_time = !timelimit; } in
+            let limits = { limits with Call_provers.limit_time = !timelimit; } in
             let callback paid st =
               callback_pa paid st;
               begin
@@ -1518,7 +1517,7 @@ later on. We do has if proof fails. *)
             in
             Debug.dprintf
               debug "[Bisect] running the prover on subtask@.";
-            schedule_proof_attempt c pn prover ~limit ~callback ~notification
+            schedule_proof_attempt c pn prover ~limits ~callback ~notification
          | _ -> assert false
       end
     in