+++ src/why3session/why3session_create.ml
@@ -77,7 +77,7 @@ let run () =
             Whyconf.Args.exit_with_usage usage_msg
   in
   let main = Whyconf.get_main config in
-  let limit =
+  let limits =
     {
       Call_provers.limit_time =
         (match !opt_timelimit with
@@ -118,7 +118,7 @@ let run () =
       match opt_trans with
       | [] -> (* When no transformations are left, add proof attempts *)
            List.iter (fun p ->
-              let _ = graft_proof_attempt session g p.Whyconf.prover ~limit in
+              let _ = graft_proof_attempt session g p.Whyconf.prover ~limits in
               ()) provers
       | (tname, targs) :: more_trans -> (* Apply tname and call recursively *)
       let new_task_list =