+++ src/why3session/why3session_update.ml
@@ -80,7 +80,7 @@ let do_action ~config ~env ~session action =
       Whyconf.filter_one_prover config filter_prover
     in
     let main = Whyconf.get_main config in
-    let limit =
+    let limits =
       {
         Call_provers.limit_time = Whyconf.timelimit main;
         Call_provers.limit_steps = 0;
@@ -88,7 +88,7 @@ let do_action ~config ~env ~session action =
       }
     in
     Why3session_lib.session_iter_proof_node_id_by_filter session f (fun pn ->
-        let _ = graft_proof_attempt session pn p.Whyconf.prover ~limit in
+        let _ = graft_proof_attempt session pn p.Whyconf.prover ~limits in
         ())
 
 let run_update () =